Moves ClipList code into Manifold.cpp
